+++ src/main/java/com/solvd/homework30nov2023/dao/IBaseDao.java
@@ -10,7 +10,7 @@ public interface IBaseDao<T> {
 
     void update(T t, Long id);
 
-    void removeById(int id);
+    void removeById(Long id);
 
     List<T> getAll();
 }
